src/backend/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
automat==20.2.0
cffi==1.14.4
channels-redis==3.2.0
channels==3.0.3
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
django-cors-headers==3.6.0
django==3.1.5
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
gunicorn==20.0.4
hyperlink==20.0.1
idna==3.1
incremental==17.5.0
msgpack==1.0.2
python-dotenv==1.0.0
psycopg2-binary
pyasn1-modules
pyasn1==0.4.8
pycparser==2.20
pyhamcrest==2.0.2
pyjwt==2.0.0
pyopenssl==20.0.1
pytz==2020.5
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
twisted
txaio==20.12.1
whitenoise==5.2.0
zope.interface==5.2.0
django_mysql
django-jsonfield